Finish the list box
I have finished the list box. For now, the list box is able to show the
coordinates and the current that user enters. However, because I don't
know the current coils number input and how the other coordinates are
stored. I don't know how to save all the results and display them. But,
I have written a for loop once everything is ready to go!